BE/refactor: auth에 있던 enum 들 globa.model로 이동
