+++ server/src/main/java/com/jh/chamgyunin/domain/auth/exception/InvalidUserProviderException.java
@@ -1,8 +1,8 @@
 package com.jh.chamgyunin.domain.auth.exception;
 
-import com.jh.chamgyunin.domain.auth.dto.UserProvider;
+import com.jh.chamgyunin.global.model.UserProvider;
 import com.jh.chamgyunin.global.error.exception.BusinessException;
-import com.jh.chamgyunin.global.error.exception.ErrorCode;
+import com.jh.chamgyunin.global.error.ErrorCode;
 
 public class InvalidUserProviderException extends BusinessException {
