+++ server/src/main/java/com/jh/chamgyunin/domain/auth/exception/SocialNotRegisteredUserException.java
@@ -1,7 +1,7 @@
 package com.jh.chamgyunin.domain.auth.exception;
 
 import com.jh.chamgyunin.global.error.exception.BusinessException;
-import com.jh.chamgyunin.global.error.exception.ErrorCode;
+import com.jh.chamgyunin.global.error.ErrorCode;
 
 public class SocialNotRegisteredUserException extends BusinessException {
     public SocialNotRegisteredUserException() {
