+++ server/src/main/java/com/jh/chamgyunin/domain/auth/exception/UserIsNotLoggedInException.java
@@ -1,7 +1,7 @@
 package com.jh.chamgyunin.domain.auth.exception;
 
 import com.jh.chamgyunin.global.error.exception.BusinessException;
-import com.jh.chamgyunin.global.error.exception.ErrorCode;
+import com.jh.chamgyunin.global.error.ErrorCode;
 
 public class UserIsNotLoggedInException extends BusinessException {
     public UserIsNotLoggedInException() {
