+++ server/src/main/java/com/jh/chamgyunin/domain/auth/service/social_userinfo/SocialUserInfoService.java
@@ -1,4 +1,4 @@
-package com.jh.chamgyunin.domain.auth.service;
+package com.jh.chamgyunin.domain.auth.service.social_userinfo;
 
 import com.jh.chamgyunin.domain.auth.dto.userinfo.SocialUserInfo;
