+++ server/src/main/java/com/jh/chamgyunin/domain/auth/service/social_userinfo/factory/SocialUserInfoServiceFactory.java
@@ -1,6 +1,8 @@
-package com.jh.chamgyunin.domain.auth.service;
+package com.jh.chamgyunin.domain.auth.service.social_userinfo.factory;
 
-import com.jh.chamgyunin.domain.auth.dto.UserProvider;
+import com.jh.chamgyunin.domain.auth.service.social_userinfo.KakaoUserInfoServiceImpl;
+import com.jh.chamgyunin.domain.auth.service.social_userinfo.SocialUserInfoService;
+import com.jh.chamgyunin.global.model.UserProvider;
 import com.jh.chamgyunin.domain.auth.exception.InvalidUserProviderException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
