+++ server/src/main/java/com/jh/chamgyunin/domain/post/exception/PostNotFoundException.java
@@ -1,7 +1,7 @@
 package com.jh.chamgyunin.domain.post.exception;
 
 import com.jh.chamgyunin.global.error.exception.BusinessException;
-import com.jh.chamgyunin.global.error.exception.ErrorCode;
+import com.jh.chamgyunin.global.error.ErrorCode;
 
 public class PostNotFoundException extends BusinessException {
     public PostNotFoundException(final Long id) {
