+++ server/src/main/java/com/jh/chamgyunin/domain/user/exception/UserAlreadyExistException.java
@@ -1,7 +1,7 @@
 package com.jh.chamgyunin.domain.user.exception;
 
 import com.jh.chamgyunin.global.error.exception.BusinessException;
-import com.jh.chamgyunin.global.error.exception.ErrorCode;
+import com.jh.chamgyunin.global.error.ErrorCode;
 
 public class UserAlreadyExistException extends BusinessException {
     public UserAlreadyExistException(final String email) {
