+++ server/src/main/java/com/jh/chamgyunin/global/error/ErrorResponse.java
@@ -1,7 +1,6 @@
 package com.jh.chamgyunin.global.error;
 
 
-import com.jh.chamgyunin.global.error.exception.ErrorCode;
 import lombok.Getter;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.method.annotation.MethodArgumentTypeMismatchException;
