+++ server/src/main/java/com/jh/chamgyunin/global/error/exception/GlobalExceptionHandler.java
@@ -1,6 +1,7 @@
 package com.jh.chamgyunin.global.error.exception;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
+import com.jh.chamgyunin.global.error.ErrorCode;
 import com.jh.chamgyunin.global.error.ErrorResponse;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
