+++ server/src/main/java/com/jh/chamgyunin/global/model/UserProvider.java
@@ -1,4 +1,4 @@
-package com.jh.chamgyunin.domain.auth.dto;
+package com.jh.chamgyunin.global.model;
 
 public enum UserProvider {
     KAKAO("kakao"),
