+++ server/src/test/java/com/jh/chamgyunin/domain/auth/api/LoginControllerTest.java
@@ -2,7 +2,7 @@ package com.jh.chamgyunin.domain.auth.api;
 
 import com.jh.chamgyunin.IntegrationTest;
 import com.jh.chamgyunin.domain.auth.dto.SocialLoginRequest;
-import com.jh.chamgyunin.domain.auth.dto.UserProvider;
+import com.jh.chamgyunin.global.model.UserProvider;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.MediaType;
